Provide smoothed, rotation-speed-sensitive adjustment of speed multiplier on LCD main status screen.
Monitor movements in the encoder wheel over a period of time; reduce the sensitivity at slow speeds (so that you can accurately change it in single steps, and not accidentally change it), while moving in larger increments for faster turns.
Make the encoder position a signed variable, since most of the logic that uses it assumes that it is (and then works around the fact that it isn't).